Removed single quotes from -File param
Single quotes causes an error when running powershell.exe: "The given path's format is not supported."

I could avoid this with double quotes, but then I'd need to figure out escaping the double quotes inside the call to cmd /c "blah"...and escaping them again inside the PS string.

Since we're already copying the file to the admin$ share and we control the filename, it should be safe to assume that there are no spaces. I hope.